# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenClaw Backend API
  version: 1.1.0
  description: |
    OpenClaw 项目后端接口文档（基于当前实现）。

    - 基础地址: http://127.0.0.1:4000
    - 管理端接口需要 admin 角色
servers:
  - url: http://127.0.0.1:4000
...
paths:
...
  /api/modules:
    get:
      tags: [Modules]
      summary: 获取当前用户开通模块
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModuleMeta'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /api/modules/{moduleKey}/history:
    get:
      tags: [Modules]
      summary: 获取模块任务历史
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ModuleKey'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModuleTaskResult'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/modules/{moduleKey}/schema:
    get:
      tags: [Modules]
      summary: 获取模块输入 schema
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ModuleKey'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleTaskSchema'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/modules/{moduleKey}/tasks:
    post:
      tags: [Modules]
      summary: 创建模块任务
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ModuleKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleTaskPayload'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleTaskResult'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /api/modules/{moduleKey}/tasks/{taskId}:
    get:
      tags: [Modules]
      summary: 获取单个模块任务
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ModuleKey'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleTaskResult'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/admin/module-factory/modules:
    get:
      tags: [Admin]
      summary: 获取已生成自定义模块
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneratedModuleListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/admin/users/{id}/modules:
    put:
      tags: [Admin]
      summary: 更新用户开通模块
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserModulesRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/admin/module/{moduleKey}/tasks:
    get:
      tags: [Admin]
      summary: 模块任务列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ModuleKey'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: featureKey
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleTaskListResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    reportAccessToken:
      type: apiKey
      in: query
      name: access
  parameters:
    ModuleKey:
      in: path
      name: moduleKey
      required: true
      schema:
        type: string
...
    TaskId:
      in: path
      name: taskId
      required: true
      schema:
        type: string
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  responses:
    ErrorResponse:
      description: 业务错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    TaskStatus:
      type: string
      enum: [review, queued, running, completed, failed]
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
...
    ModuleMeta:
      type: object
      properties:
        moduleKey:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [enterprise, personal]
        description:
          type: string
        icon:
          type: string
        status:
          type: string
        mobileSupported:
          type: boolean
    ModuleTaskResult:
      type: object
      properties:
        taskId:
          type: string
        moduleKey:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        summary:
          type: string
        updatedAt:
          type: string
          format: date-time
        reportUrl:
          type: string
        errorMessage:
          type: string
          nullable: true
    ModuleTaskPayload:
      type: object
      required: [scenario, inputText]
      properties:
        scenario:
          type: string
        inputText:
          type: string
        attachments:
          type: array
          items:
            type: string
    ModuleTaskSchemaMetric:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        unit:
          type: string
    ModuleTaskSchema:
      type: object
      properties:
        scenarios:
          type: array
          items:
            type: string
        inputHints:
          type: array
          items:
            type: string
        samplePrompt:
          type: string
        metrics:
          type: array
          items:
            $ref: '#/components/schemas/ModuleTaskSchemaMetric'
...
    GeneratedModuleItem:
      type: object
      properties:
        moduleKey:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [enterprise, personal]
        description:
          type: string
        icon:
          type: string
        status:
          type: string
          enum: [active, beta, coming_soon]
        mobileSupported:
          type: boolean
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
...
    GeneratedModuleListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GeneratedModuleItem'
    AdminTask:
      type: object
      properties:
        taskId:
          type: string
        tenantId:
          type: string
        ownerId:
          type: string
        moduleKey:
          type: string
        scenario:
          type: string
        inputText:
          type: string
        attachments:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        summary:
          type: string
        reportUrl:
          type: string
        reportFormat:
          type: string
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminTaskSummary:
      type: object
      properties:
        total:
          type: integer
        review:
          type: integer
        queued:
          type: integer
        running:
          type: integer
        completed:
          type: integer
        failed:
          type: integer
        processing:
          type: integer
        successRate:
          type: integer
...
    ModuleTaskFeatureApplied:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
        description:
          type: string
    ModuleTaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/AdminTask'
        total:
          type: integer
        summary:
          $ref: '#/components/schemas/AdminTaskSummary'
        featureApplied:
          allOf:
            - $ref: '#/components/schemas/ModuleTaskFeatureApplied'
          nullable: true
...
    UpdateUserModulesRequest:
      type: object
      required: [modules]
      properties:
        modules:
          type: array
          items:
            type: string